move test into unit tests. they were not being run before!
