+++ common/test/unit/com/thoughtworks/go/config/materials/scm/PluggableSCMMaterialTest.java
@@ -1,7 +1,8 @@
-package com.thoughtworks.go.config.materials;
+package com.thoughtworks.go.config.materials.scm;
 
 import com.thoughtworks.go.config.CaseInsensitiveString;
 import com.thoughtworks.go.config.PipelineConfig;
+import com.thoughtworks.go.config.materials.PluggableSCMMaterial;
 import com.thoughtworks.go.domain.MaterialRevision;
 import com.thoughtworks.go.domain.config.*;
 import com.thoughtworks.go.domain.materials.MatchedRevision;
